requirements.txt
nemo_toolkit["asr"]
Flask
requests
av
numpy